refactor: unify memory consolidation workers and fix UI gap\n\nUnified MemoryConsolidationWorker and ConsolidationWorker into a single entity to remove duplicate background tasks. Fixed a bug in the Knowledge UI routes pointing to an incorrect API port (8081 vs 18789), and introduced an E2E Playwright test covering knowledge document upload and deletion flows.
